+++ src/client/OrganizationAccountClient.ts
@@ -13,12 +13,15 @@ import {
   OrgTeamMemberQueryResponse,
   OrgTeamRepoQueryResponse,
   TeamRepositoryPermission,
-  OrgCollaboratorQueryResponse,
-  OrgAppQueryResponse,
   GithubResource,
-  OrgSecretQueryResponse,
-  OrgSecretRepoQueryResponse,
 } from './GraphQLClient';
+import {
+  RepoCollaboratorQueryResponse,
+  OrgAppQueryResponse,
+  SecretQueryResponse,
+  OrgSecretRepoQueryResponse,
+  RepoEnvironmentQueryResponse,
+} from './RESTClient/types';
 import {
   RepoEntity,
   ReposCompareCommitsResponseItem,
@@ -73,6 +76,17 @@ export default class OrganizationAccountClient {
     this.v4RateLimitConsumed = 0;
   }
 
+  /**
+   * Calls using the Octokit GraphQL client (API v4).
+   *
+   * GraphQL is the preferred way of accessing the GitHub API, because it calls a hierarchy of
+   * resources and only the desired properties. It is therefore more efficient.
+   * However, to find a resource in GraphQL, one must know where in the hierarchy that resource
+   * lives. Documentation at https://docs.github.com/en/graphql does not list all resources
+   * that are available through the REST API (v3).
+   *
+   */
+
   async getAccount(): Promise<OrgQueryResponse> {
     let response;
     await this.queryGraphQL('account and related resources', async () => {
@@ -87,6 +101,23 @@ export default class OrganizationAccountClient {
       return rateLimitConsumed;
     });
     return response;
+  }
+
+  async getMembers(): Promise<OrgMemberQueryResponse[]> {
+    let response;
+    await this.queryGraphQL('members', async () => {
+      const {
+        membersWithRole,
+        rateLimitConsumed,
+      } = await this.v4.fetchFromSingle(
+        GithubResource.Organization,
+        [GithubResource.OrganizationMembers],
+        { login: this.login },
+      );
+      response = membersWithRole;
+      return rateLimitConsumed;
+    });
+    return response || [];
   }
 
   async getTeams(): Promise<OrgTeamQueryResponse[]> {
@@ -180,6 +211,65 @@ export default class OrganizationAccountClient {
     return response || [];
   }
 
+  async iteratePullRequestEntities(
+    repo: RepoEntity,
+    iteratee: ResourceIteratee<PullRequest>,
+  ): Promise<PullRequestQueryResponse> {
+    if (!this.authorizedForPullRequests) {
+      this.logger.info('Account not authorized for ingesting pull requests.');
+      return { rateLimitConsumed: 0 };
+    }
+    const query = `is:pr repo:${repo.fullName}`;
+    return await this.v4.iteratePullRequests(
+      query,
+      [GithubResource.Commits, GithubResource.Reviews, GithubResource.Labels],
+      iteratee,
+    );
+  }
+
+  /**
+   * Calls using the Octokit REST client (v3 API).
+   *
+   * In general, we'd like to phase these out when possible and use GraphQL instead.
+   * We are currently using REST calls where the resource is not available through
+   * GraphQL, or possibly it is available but we don't know where to access it in the
+   * object hierarchy.
+   *
+   */
+
+  async getRepoCollaboratorsWithRest(
+    repoName: string,
+  ): Promise<RepoCollaboratorQueryResponse[]> {
+    try {
+      const repoCollaborators = await this.v3.paginate(
+        'GET /repos/{owner}/{repo}/collaborators', // https://docs.github.com/en/rest/reference/repos#list-repository-collaborators
+        {
+          owner: this.login,
+          repo: repoName,
+          per_page: 100,
+        },
+        (response) => {
+          this.logger.info('Fetched page of repo collaborators');
+          this.v3RateLimitConsumed++;
+          return response.data;
+        },
+      );
+      return repoCollaborators || [];
+    } catch (err) {
+      //this method is called for every repo in the integration, but some might have special permissions restrictions
+      //if we fail for one repo, we don't want to fail the whole collaborators step
+      this.logger.warn(
+        {
+          err: err,
+          repo: repoName,
+          endpoint: `/repos/${this.login}/${repoName}/collaborators`,
+        },
+        `Failed to retrieve collaborators for repo ${repoName}, proceeding to other repos`,
+      );
+      return [];
+    }
+  }
+
   // This is a hack to allow large github accounts to bypass a Github error. Please delete this code once that error is fixed.
   // Do not attempt to call REST version of getTeamRepositories without first calling getTeams
   async getTeamReposWithRest(
@@ -239,95 +329,7 @@ export default class OrganizationAccountClient {
     return totalTeamRepos || [];
   }
 
-  async getRepoCollaboratorsWithRest(
-    repoName: string,
-  ): Promise<OrgCollaboratorQueryResponse[]> {
-    try {
-      const repoCollaborators = await this.v3.paginate(
-        'GET /repos/{owner}/{repo}/collaborators', // https://docs.github.com/en/rest/reference/repos#list-repository-collaborators
-        {
-          owner: this.login,
-          repo: repoName,
-          per_page: 100,
-        },
-        (response) => {
-          this.logger.info('Fetched page of repo collaborators');
-          this.v3RateLimitConsumed++;
-          return response.data;
-        },
-      );
-      return repoCollaborators || [];
-    } catch (err) {
-      //this method is called for every repo in the integration, but some might have special permissions restrictions
-      //if we fail for one repo, we don't want to fail the whole collaborators step
-      this.logger.warn(
-        {
-          err: err,
-          repo: repoName,
-          endpoint: `/repos/${this.login}/${repoName}/collaborators`,
-        },
-        `Failed to retrieve collaborators for repo ${repoName}, proceeding to other repos`,
-      );
-      return [];
-    }
-  }
-
-  async getMembers(): Promise<OrgMemberQueryResponse[]> {
-    let response;
-    await this.queryGraphQL('members', async () => {
-      const {
-        membersWithRole,
-        rateLimitConsumed,
-      } = await this.v4.fetchFromSingle(
-        GithubResource.Organization,
-        [GithubResource.OrganizationMembers],
-        { login: this.login },
-      );
-      response = membersWithRole;
-      return rateLimitConsumed;
-    });
-    return response || [];
-  }
-
-  async getInstalledApps(ghsToken): Promise<OrgAppQueryResponse[]> {
-    //the endpoint needed is /orgs/${this.login}/installations
-    //when we try to call it from the Octokit v3 REST client paginate function, we get 'bad credentials'
-    //per GitHub tech support, this endpoint requires a token that starts with 'ghs'
-    //the v3 Octokit REST client .auth call returns such a token, which we pass to the
-    //v4 GraphQL client. However, the v4 GraphQL client does not appear to have access
-    //to the Apps Nodes (as far as we can currently tell), and the v3 REST client does not
-    //appear to use the 'ghs' token that it returns. When we use curl or a direct request
-    //via @octokit/request, using the ghs token, the endpoint works. Attempts to override
-    //the v3 REST client headers in the paginate function, in order to force it to use the
-    //ghs token, have been unsuccessful. After several hours of experimentation and research,
-    //the only thing that has worked in this direct call to @octokit/request.
-    //This is not ideal, since it is not a paginated call. We could build our own pagination and
-    //rate-limit aware wrapper for it, but if this endpoint is the only time we need @octokit/request,
-    //we will probably be okay without pagination and rate-limit aware features, because there are
-    //typically only going to be a few GitHub apps installed in a given organization.
-    //TODO: a more elegant solution. Possibly making our own pagination and rate-limit aware wrapper.
-    try {
-      const reply = await request(`GET /orgs/${this.login}/installations`, {
-        headers: {
-          authorization: `Bearer ${ghsToken}`,
-        },
-        type: 'private',
-      });
-      if (reply.data.installations) {
-        return reply.data.installations;
-      }
-      this.logger.warn({}, 'Found no installed GitHub apps');
-      return [];
-    } catch (err) {
-      this.logger.warn(
-        {},
-        'Error while attempting to ingest to installed GitHub apps',
-      );
-      throw new IntegrationError(err);
-    }
-  }
-
-  async getOrganizationSecrets(): Promise<OrgSecretQueryResponse[]> {
+  async getOrganizationSecrets(): Promise<SecretQueryResponse[]> {
     try {
       const orgSecrets = await this.v3.paginate(
         'GET /orgs/{org}/actions/secrets', //https://docs.github.com/en/rest/reference/actions#list-organization-secrets
@@ -378,7 +380,7 @@ export default class OrganizationAccountClient {
     }
   }
 
-  async getRepoSecrets(repoName: string): Promise<OrgSecretQueryResponse[]> {
+  async getRepoSecrets(repoName: string): Promise<SecretQueryResponse[]> {
     try {
       const repoSecrets = await this.v3.paginate(
         'GET /repos/{owner}/{repo}/actions/secrets', //https://docs.github.com/en/rest/reference/actions#list-repository-secrets
@@ -400,21 +402,107 @@ export default class OrganizationAccountClient {
     }
   }
 
-  async iteratePullRequestEntities(
-    repo: RepoEntity,
-    iteratee: ResourceIteratee<PullRequest>,
-  ): Promise<PullRequestQueryResponse> {
-    if (!this.authorizedForPullRequests) {
-      this.logger.info('Account not authorized for ingesting pull requests.');
-      return { rateLimitConsumed: 0 };
+  async getEnvironments(
+    repoName: string,
+  ): Promise<RepoEnvironmentQueryResponse[]> {
+    try {
+      const repoEnvironments = await this.v3.paginate(
+        'GET /repos/{owner}/{repo}/environments' as any, //https://docs.github.com/en/rest/reference/repos#get-all-environments
+        {
+          owner: this.login,
+          repo: repoName,
+          per_page: 100,
+        },
+        (response) => {
+          this.logger.info('Fetched page of environments for a repo');
+          this.v3RateLimitConsumed++;
+          return response.data;
+        },
+      );
+      return repoEnvironments || [];
+    } catch (err) {
+      this.logger.warn(
+        {},
+        'Error while attempting to ingest repo environments',
+      );
+      throw new IntegrationError(err);
     }
-    const query = `is:pr repo:${repo.fullName}`;
-    return await this.v4.iteratePullRequests(
-      query,
-      [GithubResource.Commits, GithubResource.Reviews, GithubResource.Labels],
-      iteratee,
-    );
   }
+
+  async getEnvSecrets(
+    repoDatabaseId: string,
+    envName: string,
+  ): Promise<SecretQueryResponse[]> {
+    try {
+      const repoSecrets = await this.v3.paginate(
+        'GET /repositories/{repository_id}/environments/{environment_name}/secrets' as any, //https://docs.github.com/en/rest/reference/actions#list-environment-secrets
+        {
+          repository_id: repoDatabaseId,
+          environment_name: envName,
+          per_page: 100,
+        },
+        (response) => {
+          this.logger.info('Fetched page of secrets for a repo environment');
+          this.v3RateLimitConsumed++;
+          return response.data;
+        },
+      );
+      return repoSecrets || [];
+    } catch (err) {
+      this.logger.warn(
+        {},
+        'Error while attempting to ingest repo environment secrets',
+      );
+      throw new IntegrationError(err);
+    }
+  }
+
+  /**
+   * One-off API calls using neither the Octokit GraphQL or REST clients for some reason.
+   */
+
+  async getInstalledApps(ghsToken): Promise<OrgAppQueryResponse[]> {
+    //the endpoint needed is /orgs/${this.login}/installations
+    //when we try to call it from the Octokit v3 REST client paginate function, we get 'bad credentials'
+    //per GitHub tech support, this endpoint requires a token that starts with 'ghs'
+    //the v3 Octokit REST client .auth call returns such a token, which we pass to the
+    //v4 GraphQL client. However, the v4 GraphQL client does not appear to have access
+    //to the Apps Nodes (as far as we can currently tell), and the v3 REST client does not
+    //appear to use the 'ghs' token that it returns. When we use curl or a direct request
+    //via @octokit/request, using the ghs token, the endpoint works. Attempts to override
+    //the v3 REST client headers in the paginate function, in order to force it to use the
+    //ghs token, have been unsuccessful. After several hours of experimentation and research,
+    //the only thing that has worked in this direct call to @octokit/request, which is the v3 REST API.
+    //It's not ideal to call that REST API without the client wrapper, because it is not a paginated call,
+    //and it does not have rate-limit and retry functions. We could build our own pagination and
+    //rate-limit aware wrapper for it, but if this endpoint is the only time we need @octokit/request,
+    //we will probably be okay without pagination and rate-limit aware features, because there are
+    //typically only going to be a few GitHub apps installed in a given organization.
+    //TODO: a more elegant solution. Possibly making our own pagination and rate-limit aware wrapper.
+    try {
+      const reply = await request(`GET /orgs/${this.login}/installations`, {
+        headers: {
+          authorization: `Bearer ${ghsToken}`,
+        },
+        type: 'private',
+      });
+      if (reply.data.installations) {
+        return reply.data.installations;
+      }
+      this.logger.warn({}, 'Found no installed GitHub apps');
+      return [];
+    } catch (err) {
+      this.logger.warn(
+        {},
+        'Error while attempting to ingest to installed GitHub apps',
+      );
+      throw new IntegrationError(err);
+    }
+  }
+
+  /**
+   * Helper functions
+   */
 
   async getComparison(
     account: string,