+++ src/sync/converters.test.ts
@@ -8,8 +8,11 @@ import {
   toRepoSecretEntity,
   toPullRequestEntity,
   createUnknownUserPrRelationship,
+  toAppEntity,
+  toEnvironmentEntity,
+  toEnvSecretEntity,
 } from './converters';
-import { UserEntity } from '../types';
+import { EnvironmentEntity, UserEntity } from '../types';
 import { PullRequest } from '../client/GraphQLClient/types';
 import {
   fixturePullRequest,
@@ -284,6 +287,99 @@ describe('toOrganizationCollaboratorEntity', () => {
   });
 });
 
+describe('toAppEntity', () => {
+  const apiResponse = {
+    id: 19232829,
+    account: {
+      login: 'Kei-Institute',
+      id: 84923503,
+      node_id: 'MDEyOk9yZ2FuaXphdGlvbjg0OTIzNTAz',
+      avatar_url: 'https://avatars.githubusercontent.com/u/84923503?v=4',
+      gravatar_id: '',
+      url: 'https://api.github.com/users/Kei-Institute',
+      html_url: 'https://github.com/Kei-Institute',
+      followers_url: 'https://api.github.com/users/Kei-Institute/followers',
+      following_url:
+        'https://api.github.com/users/Kei-Institute/following{/other_user}',
+      gists_url: 'https://api.github.com/users/Kei-Institute/gists{/gist_id}',
+      starred_url:
+        'https://api.github.com/users/Kei-Institute/starred{/owner}{/repo}',
+      subscriptions_url:
+        'https://api.github.com/users/Kei-Institute/subscriptions',
+      organizations_url: 'https://api.github.com/users/Kei-Institute/orgs',
+      repos_url: 'https://api.github.com/users/Kei-Institute/repos',
+      events_url: 'https://api.github.com/users/Kei-Institute/events{/privacy}',
+      received_events_url:
+        'https://api.github.com/users/Kei-Institute/received_events',
+      type: 'Organization',
+      site_admin: false,
+    },
+    repository_selection: 'all',
+    access_tokens_url:
+      'https://api.github.com/app/installations/19232829/access_tokens',
+    repositories_url: 'https://api.github.com/installation/repositories',
+    html_url:
+      'https://github.com/organizations/Kei-Institute/settings/installations/19232829',
+    app_id: 135931,
+    app_slug: 'kai-institute-mknoedel',
+    target_id: 84923503,
+    target_type: 'Organization',
+    permissions: {
+      members: 'read',
+      secrets: 'read',
+      metadata: 'read',
+      organization_secrets: 'read',
+      organization_administration: 'read',
+    },
+    events: [],
+    created_at: '2021-09-02T18:37:08.000Z',
+    updated_at: '2021-09-02T18:58:19.000Z',
+    single_file_name: null,
+    has_multiple_single_files: false,
+    single_file_paths: [],
+    suspended_by: null,
+    suspended_at: null,
+  };
+
+  test('properties transferred', () => {
+    const entity = toAppEntity(apiResponse as any);
+    expect(entity).toEqual({
+      _class: ['Application'],
+      _type: 'github_app',
+      _key: 'GitHubAppInstallation_19232829',
+      _rawData: [
+        {
+          name: 'default',
+          rawData: apiResponse,
+        },
+      ],
+      name: 'kai-institute-mknoedel',
+      displayName: 'kai-institute-mknoedel',
+      webLink:
+        'https://github.com/organizations/Kei-Institute/settings/installations/19232829',
+      installationId: 19232829,
+      respositorySelection: undefined,
+      appId: 135931,
+      appSlug: 'kai-institute-mknoedel',
+      targetId: 84923503,
+      targetType: 'Organization',
+      createdOn: 1630607828000,
+      updatedOn: 1630609099000,
+      events: [],
+      repositorySelection: undefined,
+      singleFileName: '',
+      hasMultipleSingleFiles: false,
+      singleFilePaths: [],
+      suspendedOn: undefined,
+      'permissions.members': 'read',
+      'permissions.secrets': 'read',
+      'permissions.metadata': 'read',
+      'permissions.organization-secrets': 'read',
+      'permissions.organization-administration': 'read',
+    });
+  });
+});
+
 describe('toOrgSecretEntity', () => {
   const apiResponse = {
     name: 'KINDA_SECRET',
@@ -348,6 +444,209 @@ describe('toRepoSecretEntity', () => {
         'https://github.com/SomeOrg/Test-repo/settings/secrets/actions/PRETTY_SECRET',
       createdOn: 1629762780000,
       updatedOn: 1629762780000,
+      visibility: 'selected',
+    });
+  });
+});
+
+describe('toEnvironmentEntity', () => {
+  const apiResponse = {
+    id: '288429400',
+    node_id: 'MDExOkVudmlyb25tZW50Mjg4NDI5NDAw',
+    name: 'Ambiente',
+    url: 'https://api.github.com/repos/SomeOrg/SomeRepo/environments/Ambiente',
+    html_url:
+      'https://github.com/SomeOrg/SomeRepo/deployments/activity_log?environments_filter=Ambiente',
+    created_at: '2021-08-23T23:53:20Z',
+    updated_at: '2021-08-23T23:53:20Z',
+    protection_rules: [
+      {
+        id: 3736,
+        node_id: 'MDQ6R2F0ZTM3MzY=',
+        type: 'wait_timer',
+        wait_timer: 30,
+      },
+      {
+        id: 3755,
+        node_id: 'MDQ6R2F0ZTM3NTU=',
+        type: 'required_reviewers',
+        reviewers: [
+          {
+            type: 'User',
+            reviewer: {
+              login: 'octocat',
+              id: 1,
+              node_id: 'MDQ6VXNlcjE=',
+              avatar_url: 'https://github.com/images/error/octocat_happy.gif',
+              gravatar_id: '',
+              url: 'https://api.github.com/users/octocat',
+              html_url: 'https://github.com/octocat',
+              followers_url: 'https://api.github.com/users/octocat/followers',
+              following_url:
+                'https://api.github.com/users/octocat/following{/other_user}',
+              gists_url: 'https://api.github.com/users/octocat/gists{/gist_id}',
+              starred_url:
+                'https://api.github.com/users/octocat/starred{/owner}{/repo}',
+              subscriptions_url:
+                'https://api.github.com/users/octocat/subscriptions',
+              organizations_url: 'https://api.github.com/users/octocat/orgs',
+              repos_url: 'https://api.github.com/users/octocat/repos',
+              events_url:
+                'https://api.github.com/users/octocat/events{/privacy}',
+              received_events_url:
+                'https://api.github.com/users/octocat/received_events',
+              type: 'User',
+              site_admin: false,
+            },
+          },
+          {
+            type: 'Team',
+            reviewer: {
+              id: 1,
+              node_id: 'MDQ6VGVhbTE=',
+              url: 'https://api.github.com/teams/1',
+              html_url: 'https://github.com/orgs/github/teams/justice-league',
+              name: 'Justice League',
+              slug: 'justice-league',
+              description: 'A great team.',
+              privacy: 'closed',
+              permission: 'admin',
+              members_url: 'https://api.github.com/teams/1/members{/member}',
+              repositories_url: 'https://api.github.com/teams/1/repos',
+              parent: null,
+            },
+          },
+        ],
+      },
+      {
+        id: 3756,
+        node_id: 'MDQ6R2F0ZTM3NTY=',
+        type: 'branch_policy',
+      },
+    ],
+    deployment_branch_policy: null,
+  };
+
+  test('properties transferred', () => {
+    const entity = toEnvironmentEntity(
+      apiResponse as any,
+      'SomeOrg',
+      'SomeRepo',
+    );
+    expect(entity).toEqual({
+      _class: ['Configuration'],
+      _type: 'github_environment',
+      _key: 'MDExOkVudmlyb25tZW50Mjg4NDI5NDAw',
+      _rawData: [
+        {
+          name: 'default',
+          rawData: apiResponse,
+        },
+      ],
+      name: 'Ambiente',
+      displayName: 'Ambiente',
+      webLink:
+        'https://github.com/SomeOrg/SomeRepo/settings/environments/288429400/edit',
+      id: '288429400',
+      nodeId: 'MDExOkVudmlyb25tZW50Mjg4NDI5NDAw',
+      url:
+        'https://api.github.com/repos/SomeOrg/SomeRepo/environments/Ambiente',
+      htmlUrl:
+        'https://github.com/SomeOrg/SomeRepo/deployments/activity_log?environments_filter=Ambiente',
+      createdOn: 1629762800000,
+      updatedOn: 1629762800000,
+      protectionRulesExist: true,
+    });
+  });
+
+  test('missing protection rules detected', () => {
+    apiResponse.protection_rules = [];
+    const entity = toEnvironmentEntity(
+      apiResponse as any,
+      'SomeOrg',
+      'SomeRepo',
+    );
+    expect(entity).toEqual({
+      _class: ['Configuration'],
+      _type: 'github_environment',
+      _key: 'MDExOkVudmlyb25tZW50Mjg4NDI5NDAw',
+      _rawData: [
+        {
+          name: 'default',
+          rawData: apiResponse,
+        },
+      ],
+      name: 'Ambiente',
+      displayName: 'Ambiente',
+      webLink:
+        'https://github.com/SomeOrg/SomeRepo/settings/environments/288429400/edit',
+      id: '288429400',
+      nodeId: 'MDExOkVudmlyb25tZW50Mjg4NDI5NDAw',
+      url:
+        'https://api.github.com/repos/SomeOrg/SomeRepo/environments/Ambiente',
+      htmlUrl:
+        'https://github.com/SomeOrg/SomeRepo/deployments/activity_log?environments_filter=Ambiente',
+      createdOn: 1629762800000,
+      updatedOn: 1629762800000,
+      protectionRulesExist: false,
+    });
+  });
+});
+
+describe('toEnvSecretEntity', () => {
+  const apiResponse = {
+    name: 'ENVIRONMENTAL_SECRET',
+    created_at: '2021-08-23T23:53:52Z',
+    updated_at: '2021-08-23T23:53:52Z',
+  };
+
+  test('properties transferred', () => {
+    const environment = {
+      _class: ['Configuration'],
+      _type: 'github_environment',
+      _key: 'MDExOkVudmlyb25tZW50Mjg4NDI5NDAw',
+      _rawData: [
+        {
+          name: 'default',
+          rawData: apiResponse,
+        },
+      ],
+      name: 'Ambiente',
+      displayName: 'Ambiente',
+      webLink:
+        'https://github.com/SomeOrg/SomeRepo/settings/environments/288429400/edit',
+      id: '288429400',
+      nodeId: 'MDExOkVudmlyb25tZW50Mjg4NDI5NDAw',
+      url:
+        'https://api.github.com/repos/SomeOrg/SomeRepo/environments/Ambiente',
+      htmlUrl:
+        'https://github.com/SomeOrg/SomeRepo/deployments/activity_log?environments_filter=Ambiente',
+      createdOn: 1629762800000,
+      updatedOn: 1629762800000,
+      protectionRulesExist: false,
+    } as EnvironmentEntity;
+    const entity = toEnvSecretEntity(
+      apiResponse as any,
+      'SomeOrg',
+      'Test-repo',
+      environment,
+    );
+    expect(entity).toEqual({
+      _class: ['Secret'],
+      _type: 'github_env_secret',
+      _key: 'GitHub_Env_Ambiente_Secret_ENVIRONMENTAL_SECRET',
+      _rawData: [
+        {
+          name: 'default',
+          rawData: apiResponse,
+        },
+      ],
+      name: 'ENVIRONMENTAL_SECRET',
+      displayName: 'ENVIRONMENTAL_SECRET',
+      webLink:
+        'https://github.com/SomeOrg/Test-repo/settings/environments/288429400/edit',
+      createdOn: 1629762832000,
+      updatedOn: 1629762832000,
       visibility: 'selected',
     });
   });